requirements.txt
appdirs==1.4.4
asgiref==3.6.0
astroid==2.6.6
attrs==21.2.0
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
black==21.7b0
bleach==6.0.0
bx-py-utils==75
cfgv==3.3.0
click==8.0.1
distlib==0.3.2
Django==4.1.6
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
drf-writable-nested==0.7.0
Faker==8.12.0
filelock==3.0.12
flake8==3.9.2
greenlet==1.1.1
icdiff==2.0.6
identify==2.2.13
importlib-metadata==6.0.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.9.3
jsonschema==4.17.3
lazy-object-proxy==1.6.0
Markdown==3.4.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.0
pathspec==0.9.0
platformdirs==2.2.0
pluggy==0.13.1
pprintpp==0.4.0
pre-commit==2.14.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.6.0
pylint==2.9.6
pyparsing==2.4.7
pyrsistent==0.19.3
pytest==6.2.4
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==5.4.1
regex==2021.8.3
six==1.16.0
SQLAlchemy==1.4.23
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
uritemplate==4.1.1
virtualenv==20.7.2
webencodings==0.5.1
wrapt==1.12.1
zipp==3.14.0
